requirements.txt
supabase
yfinance
pandas_ta
python-dotenv
requests
python-binance
pandas
xgboost
# optional for reinforcement learning/backtesting
stable-baselines3
gym